backend/requirements.txt
flask==2.2.3
werkzeug==2.2.3
flask-cors==3.0.10
flask-limiter==3.5.0
python-dotenv==1.0.0
openai>=1.6.1,<2.0.0
langchain-core>=0.1.0,<0.2.0
langchain>=0.0.267,<0.1.0
langchain-community>=0.0.6,<0.1.0
langchain-openai>=0.0.2,<0.1.0
tiktoken>=0.5.2,<0.6.0
requests==2.31.0
numpy>=1.26.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
fpdf==1.7.2
pytest==7.4.0
pytest-mock==3.11.1 